refactor(test): helper_test.go as a useful example
